Consolidate test JSX intrinsics into one shared declaration
Three test files each augmented JSX.IntrinsicElements with slightly
different shapes, which TS merges program-wide and rejects (TS2717).
One shared test-intrinsics.d.ts now declares them once.
